+++ tests/sentry/api/path_params/resources/test_doc_integration_slug.py
