+++ tests/sentry/api/path_params/resources/test_id_or_slug_path_params_mixin.py
