+++ tests/sentry/api/path_params/resources/test_organization_slug.py
