+++ tests/sentry/api/path_params/resources/test_sentry_app_slug.py
